Update to syn 0.12
In tracking down the cause of rust-lang/rust#48223 I found that the `syn` crate
was compiled twice and with different features than the ones that Cargo
activated. It turns out that the RLS pulls in this crate which depends on `syn`,
and this crate enabled the `full` feature in `syn` where a standalone build of
Cargo didn't activate such a feature. Consequently this means that when Cargo
was compiled for the RLS and for itself it was compiled twice!

I initially started out by removing the dependency on the `full` feature in this
crate (hurray faster compiles too!) but ended up updating to syn 0.12 as well
which is the updated version that works better with spans and such.

I've also done some reorganization of the tests cases to hopefully be more
amenable to CI as well, but let me know if any of it looks odd!